Separated Pdf into a TablePdf
